+++ src/modules/proxy/assets/proxy-stop.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+APPNAME=<%= appName %>
+
+sudo docker rm -f $APPNAME || :
+docker network disconnect bridge -f $APPNAME || :
+
+sudo docker rm -f $APPNAME-letsencrypt || :